+++ src/components/misc/Skill/ChildSkillItem.js
@@ -9,6 +9,7 @@ import { ConfigContext } from '../../../context/ConfigContext';
 import { deleteSkill,completeSkill} from '../../api/SkillAPI';
 import SkillDescription from './SkillDescription';
 import AddChildSkillForm from './AddChildSkillForm';
+import { ActiveContext } from '../../../context/ActiveContext';
 
 
 const ChildSkillItem = (props) => {
@@ -19,15 +20,17 @@ const ChildSkillItem = (props) => {
     const {config} = useContext(ConfigContext);
 	const [showChildSkills, setShowChildSkills] = useState(false);
 	const [showAddSkill,setShowAddSkill] = useState(false);
-
+	const {showActive} = useContext(ActiveContext);
 	const onDelete = async() =>{
-		await deleteSkill(config,'Bearer '+user.accessToken,props.record.id)
-		await props.refreshFunction(config,'Bearer '+user.accessToken)
+		if (window.confirm('Are you sure you wish to delete this item?')){
+			await deleteSkill(config,'Bearer '+user.accessToken,props.record.id)
+			await props.refreshFunction(config,'Bearer '+user.accessToken,props.record.skillTypeName,showActive)
+		}
 	}
 
 	const onRefresh = async() =>{
 		setShowAddSkill(false);
-		await props.refreshFunction(config,'Bearer '+user.accessToken)
+		await props.refreshFunction(config,'Bearer '+user.accessToken,props.record.skillTypeName,showActive)
 	}
 
 	const onShowDescription = async() =>{
